Speed up processing feeds
Add a get_items() generator to only parse an item when needed
This allows for quicker checking of podcasts to only get the most
recently published episode.

This change is backwards compatible with the existing usage, but
when pre_set_items is disabled, the dict will not include the list
of items. All other functionality remains the same